Move redundant config stuff to shared module again
